Refactor: move mesh logic to useMesh composable.
